Fix functions and function names when creating LinearModelAlgorithmResult
Functions were inserted in the same order as in original basis.
Thus result was right if they appear in the same order in output,
but wrong otherwise.

This may happen with 'forward' or 'both' directions, but not 'backward'.
As we were able to compare our results with R step method only for
'backward' direction, this bug had not been detected.